Remove method.cpp
